feat(user): separate data by user
Separate data by user_id.

+++ app/Controllers/Http/CategoryController.js
@@ -19,43 +19,52 @@ const validateCategory = async (attributes) => {
 }
 
 class CategoryController {
-  async index () {
+  async index ({ auth }) {
     const categories = await Category
       .query()
+      .where('user_id', auth.user.id)
       .orderBy('name', 'asc')
       .fetch()
 
     return { categories }
   }
 
-  async store ({ request, response }) {
+  async store ({ request, response, auth }) {
     try {
       await validateCategory(request.all())
-      const category = await Category.create(request.only(['name']))
+      const { name } = request.all()
+      const { id } = auth.user
+      const category = await Category.create({ name, user_id: id })
       return category
     } catch (error) {
       return response.status(400).send(error.message)
     }
   }
 
-  async update ({ params, request, response }) {
+  async update ({ params, request, response, auth }) {
     try {
       await validateCategory(request.all())
       const { name } = request.all()
       const category = await Category.find(params.id)
-      category.name = name
-      await category.save()
+      if (category.user_id === auth.user.id) {
+        category.name = name
+        await category.save()
+        return category
+      }
 
-      return category
+      return response.status(403).send('Forbidden')
     } catch(error) {
       return response.status(400).send(error.message)
     }
   }
 
-  async destroy ({ params }) {
+  async destroy ({ params, response, auth }) {
     const category = await Category.find(params.id)
     if (!category) {
       return response.status(404).send('category not found')
+    }
+    if (category.user_id !== auth.user.id) {
+      return response.status(403).send('Forbidden')
     }
     await category.delete()
